pyproject.toml
[tool.poetry]
name = "pendle-telegram-loop-scout"
version = "0.1.0"
description = "Telegram bot that detects looping opportunities on Pendle Finance"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
python-telegram-bot = {version = ">=21.0", extras = ["job-queue"]}
httpx = ">=0.27"
python-dotenv = ">=1.0"
langgraph = ">=0.2"

[tool.poetry.group.notebook.dependencies]
jupyter = ">=1.0"
ipykernel = ">=6.0"

[tool.poetry.scripts]
loop-scout = "telegram_bot.bot:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"